feat: 種族の恐怖耐性を opt-in 反映し共通述語を util へ昇格（提案C1第4弾）
種族の恐怖耐性 TR_RES_FEAR を反映。全恐怖源が通る中央ゲート
effect_damage_piles_fear (effect-monster.cpp) の do_fear→恐怖 変換を、NO_FEAR と
同経路へ || target_race_resists_element(em_ptr, TR_RES_FEAR) を OR-in して無効化。
ダメージ属性でない状態異常耐性を反映した最初の例。

- 共通述語 target_race_resists_element(EffectMonster*, tr_type) を
  effect-monster-resist-hurt.cpp の匿名 namespace から effect-monster-util.{h,cpp}
  へ昇格（属性ダメージ／状態異常の両 TU から使う共通述語に）。
- 同じ opt-in フラグ applies_player_race_resistances を使用（新フラグなし）。
  既定 false のままバランス不変。CLAUDE.md / roadmap 整備。
  フルビルド(g++ -O3 -Werror)で検証済。
